Swagger support for authorisations, fix swagger-codegen
This involves moving the entire api to /sapi due to
stupid framework restrictions and to avoid calling the default path "/",
which caused problems with Swagger code generators.

Authorisations makes the Swagger UI display a button for logging in, and
will additionally document the used authorisations. However, this relies
on hard-coding certain things (notably the login URL) in
`apis/common/auth.py`, as authorisations needs to be available when the
Api object is instantiated, very early in the initialisation process.